python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lansi-arm-sdk"
version = "1.0.0"
description = "Lansi 7-DOF Robot Arm SDK - Professional robotics development toolkit"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Lansi Robotics", email = "support@lansi-robotics.com"}
]
requires-python = ">=3.9"
dependencies = [
    "python-can>=4.0.0",
    "numpy>=1.20.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
docs = [
    "mkdocs>=1.4.0",
    "mkdocs-material>=9.0.0",
]
scipy = [
    "scipy>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/lansi-robotics/lansi_arm_sdk"
Repository = "https://github.com/lansi-robotics/lansi_arm_sdk"
Issues = "https://github.com/lansi-robotics/lansi_arm_sdk/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["lansi_arm*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
